Fix demo page flash on hydration: match SSR to interactive render
The demo swaps SSR `children` for the client `DemoRouter` after hydration
(DemoLayoutContent). The two rendered deliberately different markup, so the
swap flashed:

- Entry list cards used `bg-zinc-50/zinc-800` + `font-medium`, while the
  interactive `EntryListItem` uses the `bg-surface` token + `font-semibold`.
- Entry pages SSR'd a bare `EntryArticle` with no back link, no action
  buttons, no per-article CTA, and no reader font/size, all of which popped
  in when DemoRouter took over.

Fixes:

- Extract `DemoArticleView` (the full article view with all slots and reader
  text styles) and render it from BOTH the SSR route pages and DemoRouter, so
  the pre- and post-hydration article markup is identical. Live read/starred
  state comes from DemoStateContext (defaults on the server, matching the
  client's first render — hydrates without mismatch).
- Extract `getItemClasses` into a plain module (`entryItemClasses.ts`) so the
  crawlable SSR list (`DemoEntryListSSR`) can reuse the exact same card
  styling as `EntryListItem`. It can't reuse the interactive item directly:
  that renders `<article role="button">` with nested toggle <button>s, which
  are invalid and uncrawlable inside an `<a>`. Next.js also forbids calling a
  `"use client"` export during server render, hence the split.
