+++ scripts/raw-color-baseline.json
@@ -4,15 +4,11 @@
   "src/app/(app)/AppLayoutContent.tsx\tdark:active:bg-zinc-700",
   "src/app/admin/layout.tsx\tdark:bg-zinc-950",
   "src/app/complete-signup/page.tsx\tdark:bg-zinc-800",
-  "src/app/demo/DemoEntryListSSR.tsx\tactive:bg-zinc-200",
-  "src/app/demo/DemoEntryListSSR.tsx\tbg-zinc-50",
-  "src/app/demo/DemoEntryListSSR.tsx\tdark:active:bg-zinc-700",
-  "src/app/demo/DemoEntryListSSR.tsx\tdark:bg-zinc-800",
+  "src/app/demo/DemoArticleView.tsx\tactive:bg-zinc-200",
+  "src/app/demo/DemoArticleView.tsx\tdark:active:bg-zinc-700",
   "src/app/demo/DemoLayoutContent.tsx\tactive:bg-zinc-100",
   "src/app/demo/DemoLayoutContent.tsx\tactive:bg-zinc-200",
   "src/app/demo/DemoLayoutContent.tsx\tdark:active:bg-zinc-700",
-  "src/app/demo/DemoRouter.tsx\tactive:bg-zinc-200",
-  "src/app/demo/DemoRouter.tsx\tdark:active:bg-zinc-700",
   "src/app/extension/save/client.tsx\tborder-zinc-900",
   "src/app/extension/save/client.tsx\tdark:border-zinc-100",
   "src/app/oauth/consent/ConsentForm.tsx\tdark:bg-zinc-800",
@@ -41,13 +37,13 @@
   "src/components/entries/EntryContentRenderer.tsx\tprose-code:text-zinc-800",
   "src/components/entries/EntryContentRenderer.tsx\tprose-headings:text-zinc-900",
   "src/components/entries/EntryContentRenderer.tsx\tprose-pre:bg-zinc-100",
-  "src/components/entries/EntryListItem.tsx\tactive:bg-zinc-100",
-  "src/components/entries/EntryListItem.tsx\tborder-zinc-500",
-  "src/components/entries/EntryListItem.tsx\tdark:active:bg-zinc-700",
   "src/components/entries/EntryListItem.tsx\tdark:text-zinc-600",
-  "src/components/entries/EntryListItem.tsx\tepaper:border-zinc-500",
   "src/components/entries/EntryListItem.tsx\ttext-zinc-300",
   "src/components/entries/StickyEntryControls.tsx\tdark:bg-zinc-900/95",
+  "src/components/entries/entryItemClasses.ts\tactive:bg-zinc-100",
+  "src/components/entries/entryItemClasses.ts\tborder-zinc-500",
+  "src/components/entries/entryItemClasses.ts\tdark:active:bg-zinc-700",
+  "src/components/entries/entryItemClasses.ts\tepaper:border-zinc-500",
   "src/components/feeds/EditSubscriptionDialog.tsx\tdark:bg-zinc-800",
   "src/components/feeds/EditSubscriptionDialog.tsx\tdark:hover:bg-zinc-700",
   "src/components/feeds/EditSubscriptionDialog.tsx\tdark:hover:border-zinc-500",
