+++ src/app/demo/all/page.tsx
@@ -8,9 +8,9 @@
  */
 
 import { type Metadata } from "next";
-import { EntryArticle } from "@/components/entries/EntryArticle";
 import { pageOpenGraph } from "@/lib/metadata";
-import { getDemoEntry, getDemoEntryArticleProps, DEMO_ENTRIES, sortNewestFirst } from "../data";
+import { getDemoEntry, DEMO_ENTRIES, sortNewestFirst } from "../data";
+import { DemoArticleView } from "../DemoArticleView";
 import { DemoEntryListSSR } from "../DemoEntryListSSR";
 
 interface Props {
@@ -39,7 +39,7 @@ export default async function DemoAllPage({ searchParams }: Props) {
   const entry = entryId ? getDemoEntry(entryId) : undefined;
 
   if (entry) {
-    return <EntryArticle {...getDemoEntryArticleProps(entry)} />;
+    return <DemoArticleView entry={entry} backHref="/demo/all" />;
   }
 
   return (
